requirements.txt
streamlit
groq
python-dotenv
reportlab
pandas
numpy